+++ include/connection.h
@@ -1,5 +1,6 @@
 #ifndef CONNECTION_H
 #define CONNECTION_H
+#include <stdint.h>
 #include "buffer.h"
 #include "dy_char.h"
 #include "network.h"
@@ -26,7 +27,7 @@ struct read_context_s{
   char_t *command;
   char_t *key;
   char_t *num;
-  int last_bytes;
+  intptr_t last_bytes;
   enum PARSE_STATUS read_process;
 };
 
@@ -35,19 +36,19 @@ typedef struct read_context_s read_context_t;
 struct write_context_s{
   queue_t *w_queue;
   item_t *w_data;
-  int w_index;
+  intptr_t w_index;
 };
 
 typedef struct write_context_s write_context_t;
 
 struct connection_s{
-  int fd;
+  intptr_t fd;
   buffer_t *rbuf;
   buffer_t *wbuf;
   event_context_t *ec;
   write_context_t *wc;
   read_context_t *rc; 
-  int events; 
+  intptr_t events; 
 };
 
 typedef struct connection_s connection_t;
@@ -55,7 +56,7 @@ typedef struct connection_s connection_t;
 
 void reset_read_context(read_context_t *rc);
 
-connection_t* init_connection(int fd);
+connection_t* init_connection(intptr_t fd);
 
 void cancel_connection(connection_t *conn);
 #endif
