+++ include/statistics.h
@@ -1,19 +1,20 @@
+#include <stdint.h>
 #include "hash.h"
 #include "memory.h"
 struct queue_stat_s{
 	char *name;
-	unsigned long size;
-	unsigned long bytes;
+	intptr_t size;
+	intptr_t bytes;
 };
 
 typedef struct queue_stat_s queue_stat_t;
 
 struct stat_s{
-	unsigned long total_items;
-	int current_connections;
-	unsigned long *small_buddy_stats;
-	unsigned long *big_buddy_stats;
-	unsigned long direct_size;
+	intptr_t total_items;
+	intptr_t current_connections;
+	intptr_t *small_buddy_stats;
+	intptr_t *big_buddy_stats;
+	intptr_t direct_size;
 	hash_t *queue_data_hash;
 };
 
@@ -25,7 +26,7 @@ void init_queue_stat();
 
 void start_queue_stat(char *key);
 
-void increase_queue_stat(char *key,int delta_bytes);
+void increase_queue_stat(char *key,intptr_t delta_bytes);
 
-void decrease_queue_stat(char *key,int delta_bytes);
+void decrease_queue_stat(char *key,intptr_t delta_bytes);
